Move cnl::scale into a separate header file
